FIX: Finish refactor, tests pass
